delete Useless dependencies, fix Small bugs, add New files
