+++ app.js
@@ -4,25 +4,15 @@ const bodyParser = require('body-parser');
 const cookies = require('cookie-parser');
 const { errors } = require('celebrate');
 const helmet = require('helmet');
-const rateLimit = require('express-rate-limit');
 
-const { login, createUser, signout } = require('./controllers/users');
-const { signupValidation } = require('./middlewares/signupValidation');
-const { signinValidation } = require('./middlewares/signinValidation');
-const { auth } = require('./middlewares/auth');
 const errorHandler = require('./middlewares/error-handler');
-const NotFoundError = require('./utils/not-found-error');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
+const router = require('./routes/index');
+const { mongoUrl } = require('./utils/config');
+const limiter = require('./middlewares/limiter');
 
 const { PORT = 3000 } = process.env;
 const app = express();
-
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-  max: 100,
-  standardHeaders: true,
-  legacyHeaders: false,
-});
 
 app.use(helmet());
 app.use(cookies());
@@ -31,7 +21,7 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(requestLogger);
 
-mongoose.connect('mongodb://127.0.0.1:27017/bitfilmsdb', { useNewUrlParser: true })
+mongoose.connect(mongoUrl, { useNewUrlParser: true })
   .then(() => {
     console.log('Успешно установлена связь с MongoDB');
   })
@@ -39,18 +29,9 @@ mongoose.connect('mongodb://127.0.0.1:27017/bitfilmsdb', { useNewUrlParser: true
     console.log(`Произошла ошибка при установлении связи с MongoDB: ${error}`);
   });
 
-app.use('/users', auth, require('./routes/users'));
-app.use('/movies', auth, require('./routes/movies'));
-
-app.post('/signup', signupValidation, createUser);
-app.post('/signin', signinValidation, login);
-app.post('/signout', auth, signout);
+app.use(router);
 
 app.use(errorLogger);
-
-app.use('*', (req, res, next) => {
-  next(new NotFoundError('Страница не найдена'));
-});
 
 app.use(errors());
