+++ middlewares/signupValidation.js
@@ -7,7 +7,7 @@ module.exports.signupValidation = celebrate({
       .max(30)
       .email()
       .required(),
-    password: Joi.string().min(2).max(30).required(),
+    password: Joi.string().min(6).max(30).required(),
     name: Joi.string().min(2).max(30).required(),
   }),
 });
