+++ utils/config.js
@@ -1,0 +1,7 @@
+const mongoUrl = 'mongodb://127.0.0.1:27017/bitfilmsdb';
+const jwtSecret = 'dev-secret';
+
+module.exports = {
+  mongoUrl,
+  jwtSecret,
+};
