--- utils/createdStatus.js
@@ -1,5 +1,0 @@
-const CREATED_STATUS = 201;
-
-module.exports = {
-  CREATED_STATUS,
-};
